third_party/blink/web_tests/fast/dom/TreeWalker/previousNodeLastChildReject-expected.txt

Test for a specific problem with previousNode that failed in older versions of WebKit.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS walker.firstChild(); walker.currentNode.id is 'A1'
PASS walker.nextNode(); walker.currentNode.id is 'B1'
PASS walker.nextNode(); walker.currentNode.id is 'C1'
PASS walker.nextNode(); walker.currentNode.id is 'B2'
PASS walker.previousNode(); walker.currentNode.id is 'C1'
PASS successfullyParsed is true

TEST COMPLETE
